+++ app/lib/data/services/clerk_auth_service.dart
@@ -1,2 +1,3 @@
 export 'clerk_auth_service_stub.dart'
-    if (dart.library.html) 'clerk_auth_service_web.dart';
+    if (dart.library.html) 'clerk_auth_service_web.dart'
+    if (dart.library.io) 'clerk_auth_service_android.dart';